requirements.txt
streamlit
transformers
torch
pandas
scikit-learn
spacy
python-dotenv
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
PyPDF2
requests
beautifulsoup4
google-search-results
fastapi
uvicorn 